requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
langgraph==0.2.0
langchain==0.2.0
langchain-groq==0.1.5
groq==0.9.0
PyGithub==2.3.0
praw==7.7.1
requests==2.32.3
spacy==3.8.3
sentence-transformers==3.0.1
scikit-learn==1.5.0
shap==0.45.1
numpy==1.26.4
pandas==2.2.2
joblib==1.4.2
pdfplumber==0.11.0
langdetect==1.0.9
python-dotenv==1.0.1
httpx==0.27.0